+ refactor mysql | SingularTable setting, ParseTime to bool, + user | password check, + fix redis hmset id, + fix RefreshToken was overriding AuthToken, + fix response ReturnCode changed to int16, + register, login | better error handling

+++ models/response.go
@@ -7,7 +7,7 @@ import (
 
 // Response used for displaying errors from handlers to end user in a json format
 type Response struct {
-	ReturnCode int8
+	ReturnCode int16
 	Message    string
 }
